Remove unneccessary test combinations.
When the global stack shapes are at the top level we get many unneeded test combinations. Move them into the scope in which they are needed.
